+++ functions/convert_video.sh
@@ -10,11 +10,20 @@ HANDBRAKEPRESET="High Profile"
 # generate date and time variabel for the output folders
 NOW=$(date +"%Y-%m-%d %H:%M")
 
-for filename in *.*; do
-	# the handbrake convert code
-	# added nice to which starts the converting process in lowes priotiry so you can still use your computer
-	# HandBrakeCLI -i "$filename" -o "${filename%.*}_converted.m4v" --preset="High Profile"
-	nice -n 19 HandBrakeCLI -i "$filename" -o "${filename%.*}_converted.m4v" --preset="Fast 1080p30"
-	mv "${filename%.*}_converted.m4v" ../out/
-	rm "$filename"
-done
+if hash HandBrakeCLI 2>/dev/null; then
+	for filename in *.*; do
+		# the handbrake convert code
+		# added nice to which starts the converting process in lowes priority so you can still use your computer
+		# HandBrakeCLI -i "$filename" -o "${filename%.*}_converted.m4v" --preset="High Profile"
+		nice -n 19 HandBrakeCLI -i "$filename" -o "${filename%.*}_converted.m4v" --preset="Fast 1080p30"
+		mv "${filename%.*}_converted.m4v" ../out/
+		rm "$filename"
+	done
+	
+else
+	echo ERROR: Could not find HandBrakeCLI >&2
+	echo HandBrakeCLI needs to installed and in standardpath for convert_video to work >&2
+fi
+
+
+
